move tests to new module
